FacturIA/requirements.txt
flask
pandas
sqlalchemy
psycopg2-binary
python-dotenv
PyPDF2
openai
